requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.10.0
python-dotenv==1.0.0
openai==1.84.0
autogen-core==0.6.1
autogen-agentchat==0.6.1
autogen-ext[azure]==0.6.1
azure-core>=1.24.0
azure-identity>=1.12.0
aiohttp>=3.8.0
tiktoken==0.7.0
# Celery for task queue management
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0
# State machine for workflow management
transitions==0.9.0
# YAML configuration support
PyYAML==6.0.1
# Enhanced monitoring and metrics
prometheus-client==0.19.0
